# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: DataFinOps Commercial V1 API
  version: 1.0.0
  description: Tenant-scoped TxLINE inventory, optimization, governance, non-custodial transaction, entitlement, and savings API.
servers:
  - url: https://api.sandbox.datafinops.example/v1
    description: Deterministic sandbox contract host
security:
  - bearerAuth: []
...
paths:
...
  /requirements:
    post:
      tags: [Requirements]
      summary: Create an immutable hard and soft requirement version
      operationId: createRequirementVersion
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequirementSet' }
      responses:
        '201':
          description: Immutable requirement version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VersionReference' }
...
  /proposals:
    post:
      tags: [Proposals]
      summary: Create a policy-evaluated proposal from a verified candidate
      operationId: createProposal
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [recommendationId, candidateHash, effectiveAt]
              properties:
                recommendationId: { type: string, format: uuid }
                candidateHash: { $ref: '#/components/schemas/Sha256' }
                effectiveAt: { type: string, format: date-time }
      responses:
        '201':
          description: Proposal version and policy outcome
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Proposal' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
  responses:
    Conflict:
      description: Immutable version, expiry, approval, or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Sha256:
      type: string
      pattern: '^[a-f0-9]{64}$'
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        correlationId: { type: string }
        reasonCodes:
          type: array
          items: { type: string }
...
    VersionReference:
      type: object
      additionalProperties: false
      required: [id, version, hash, createdAt]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        hash: { $ref: '#/components/schemas/Sha256' }
        createdAt: { type: string, format: date-time }
...
    RequirementSet:
      type: object
      additionalProperties: false
      required: [workspaceId, periodStart, periodEnd, requirements]
      properties:
        workspaceId: { type: string, format: uuid }
        periodStart: { type: string, format: date-time }
        periodEnd: { type: string, format: date-time }
        requirements:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [kind, hardness, operator, value, ownerRef]
            properties:
              kind: { type: string, enum: [LEAGUE, FIXTURE, LATENCY, REGION, BUDGET, EFFECTIVE_DATE, BLACKOUT, HEADROOM] }
              hardness: { type: string, enum: [HARD, SOFT] }
              operator: { type: string, enum: [INCLUDE, EXCLUDE, MAX, MIN, EQUALS, BEFORE, AFTER] }
              value: {}
              ownerRef: { type: string, minLength: 1, maxLength: 128 }
...
    Proposal:
      type: object
      additionalProperties: false
      required: [id, version, proposalHash, snapshotHash, state, policyResult, expiresAt, approvalGraph]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        proposalHash: { $ref: '#/components/schemas/Sha256' }
        snapshotHash: { $ref: '#/components/schemas/Sha256' }
        state: { type: string, enum: [DRAFT, BLOCKED, AWAITING_APPROVAL, APPROVED, READY_TO_SIGN, EXPIRED, INVALIDATED] }
        policyResult: { type: string, enum: [ELIGIBLE, APPROVAL_REQUIRED, BLOCKED] }
        expiresAt: { type: string, format: date-time }
        approvalGraph:
          type: array
          items:
            type: object
            required: [role, order, status]
            properties:
              role: { type: string }
              order: { type: integer, minimum: 1 }
              status: { type: string, enum: [PENDING, APPROVED, REJECTED, EXPIRED] }
...